requirements.txt
asgiref==3.7.2
Django==4.2.9
django-environ==0.11.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
